requirements.txt
addict==2.4.0
asynctest==0.13.0
blinker==1.4
boto3==1.21.3
botocore==1.24.3
brotlipy==0.7.0
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.32
easydict==1.10
einops==0.4.0
fonttools==4.33.3
google-auth-oauthlib==0.4.1
imageio==2.19.3
importlib-metadata==4.11.1
inplace-abn==1.1.0
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-service==2.4.0
munch==2.5.0
numpy==1.21.6
oauthlib==3.1.0
opencv-python==4.5.5.62
pandas==1.3.4
patsy==0.5.3
Pillow==9.0.1
pretrainedmodels==0.7.4
protobuf==3.20.1
ptflops==0.6.9
pthflops==0.4.2
pyasn1-modules==0.2.8
pycocotools==2.0.4
PyJWT==1.7.1
pytorch-transformers==1.2.0
PyYAML==6.0
randaugment==1.0.2
regex==2022.1.18
requests-oauthlib==1.3.0
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
sklearn==0.0
statsmodels==0.13.5
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.5.4
torch==1.12.0
torchaudio==0.12.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.13.0
tqdm==4.62.3
visualize==0.5.0
warmup-scheduler==0.3
yapf==0.32.0
